+++ klibs/KLEyeTracking/KLEyeLink.py
@@ -1,8 +1,8 @@
 # -*- coding: utf-8 -*-
 __author__ = 'Jonathan Mulle & Austin Hurst'
 
+import os
 import time
-from os.path import join
 from klibs.KLEyeTracking import PYLINK_AVAILABLE
 
 from klibs.KLExceptions import TrialException, EyeTrackerError
@@ -13,20 +13,22 @@ from klibs.KLConstants import (EL_LEFT_EYE, EL_RIGHT_EYE, EL_BOTH_EYES, EL_NO_EY
 	EL_ALL_EVENTS, EL_TRUE, EL_FALSE,
 	TK_S, TK_MS, CIRCLE_BOUNDARY, RECT_BOUNDARY)
 from klibs import P
-from klibs.KLUtilities import full_trace, iterable, hide_mouse_cursor, mouse_pos, now
-from klibs.KLUtilities import colored_stdout as cso
+from klibs.KLInternal import full_trace, valid_coords, now, hide_stderr
+from klibs.KLInternal import colored_stdout as cso
 from klibs.KLUserInterface import ui_request
+from klibs.KLUtilities import hide_mouse_cursor
 from klibs.KLGraphics import blit, fill, flip, clear
 from klibs.KLGraphics.KLDraw import drift_correct_target
 from klibs.KLEyeTracking.KLEyeTracker import EyeTracker
 
 if PYLINK_AVAILABLE:
-	from pylink import (openGraphicsEx, flushGetkeyQueue, pumpDelay,
-		beginRealTimeMode, endRealTimeMode, msecDelay)
-	from pylink import EyeLink as BaseEyeLink
+	with hide_stderr(macos_only=True):
+		from pylink import (
+			openGraphicsEx, beginRealTimeMode, endRealTimeMode, flushGetkeyQueue,
+			pumpDelay, msecDelay
+		)
+		from pylink import EyeLink as BaseEyeLink
 	from .KLCustomEyeLinkDisplay import ELCustomDisplay
-	cso("<green_d>(Note: if a bunch of SDL errors were just reported, this was expected, "
-		"do not be alarmed!)</green_d>")
 
 class EyeLink(BaseEyeLink, EyeTracker):
 	"""A connection to an SR Research EyeLink eye tracker, providing a friendly interface to the
@@ -154,8 +156,11 @@ class EyeLink(BaseEyeLink, EyeTracker):
                 subfolder of the eye tracker data directory ('ExpAssets/EDF'). Defaults to False.
 
 		"""
-		# Determine whether EDF should go to 'incomplete' subfolder or not
+		# Determine destination path for EDF (creating parent folder if needed)
 		edf_dir = P.incomplete_edf_dir if incomplete else P.edf_dir
+		if not os.path.isdir(edf_dir):
+			os.makedirs(edf_dir)
+		edf_path = os.path.join(edf_dir, self.edf_filename)
 
 		self._quitting = True
 		if self.isRecording() == 0:
@@ -164,7 +169,7 @@ class EyeLink(BaseEyeLink, EyeTracker):
 		self.setOfflineMode()
 		msecDelay(500)
 		self.closeDataFile()
-		self.receiveDataFile(self.edf_filename, join(edf_dir, self.edf_filename))
+		self.receiveDataFile(self.edf_filename, edf_path)
 		return self.close()
 
 
@@ -260,7 +265,7 @@ class EyeLink(BaseEyeLink, EyeTracker):
 		target = drift_correct_target() if target is None else target
 		draw_target = EL_TRUE if draw_target in [EL_TRUE, True] else EL_FALSE
 		location = P.screen_c if location is None else location
-		if not iterable(location):
+		if not valid_coords(location):
 			raise ValueError("'location' must be a pair of (x,y) pixel coordinates.")
 
 		try: